+++ nodejs-server/service/ReferralsService.js
@@ -75,6 +75,79 @@ exports.createReferral = function(body) {
 
 
 /**
+ * Fetch a referral by id including its referer, referee, referer code, link, status, and method
+ *
+ * referralId Long ID of a referral to return
+ * returns referral
+ **/
+exports.getReferral = function(referralId) {
+  return new Promise(function(resolve, reject) {
+    var examples = {};
+    examples['application/json'] = [ {
+  "name" : "textSent",
+  "id" : "1",
+  "referer" : "10",
+  "referee" : "11",
+  "code" : "XFCDOB",
+  "link" : "https://cartoncaps.link/abfilefa90p?referral_code=XFCDOB",
+  "method" : "text",
+  "status" : "Sent"
+}, {
+  "name" : "emailReceived",
+  "id" : "2",
+  "referer" : "10",
+  "referee" : "21",
+  "code" : "XFCDOB",
+  "link" : "https://cartoncaps.link/abfilefa90p?referral_code=XFCDOB",
+  "status" : "Received",
+  "method" : "Email"
+}, {
+  "name" : "shareComplete",
+  "id" : "3",
+  "referer" : "30",
+  "referee" : "31",
+  "code" : "MRKCVW",
+  "link" : "https://cartoncaps.link/abfilefa90p?referral_code=MRKCVW",
+  "status" : "Complete",
+  "method" : "Share"
+}, {
+  "name" : "textReceived",
+  "id" : "4",
+  "referer" : "40",
+  "referee" : "41",
+  "code" : "LJEODN",
+  "link" : "https://cartoncaps.link/abfilefa90p?referral_code=LJEODN",
+  "status" : "Received",
+  "method" : "Text"
+}, {
+  "name" : "emailComplete",
+  "id" : "5",
+  "referer" : "50",
+  "referee" : "51",
+  "code" : "YGFVHG",
+  "link" : "https://cartoncaps.link/abfilefa90p?referral_code=YGFVHG",
+  "status" : "Complete",
+  "method" : "Email"
+}, {
+  "name" : "shareSent",
+  "id" : "6",
+  "referer" : "60",
+  "referee" : "61",
+  "code" : "WNQLRP",
+  "link" : "https://cartoncaps.link/abfilefa90p?referral_code=WNQLRP",
+  "status" : "Sent",
+  "method" : "Share"
+} ];
+    if (Object.keys(examples).length > 0) {
+      resolve(examples[Object.keys(examples)[0]]);
+    } else {
+      resolve();
+    }
+  });
+}
+
+
+/**
  * List all referrals
  *
  * limit Integer Limits the number of referrals per page (optional)
@@ -193,6 +266,28 @@ exports.getReferrals = function(limit,offset) {
   "status" : "Sent",
   "method" : "Share"
 } ] ];
+    if (Object.keys(examples).length > 0) {
+      resolve(examples[Object.keys(examples)[0]]);
+    } else {
+      resolve();
+    }
+  });
+}
+
+
+/**
+ * Updates a referral in system
+ *
+ * body Referrals_referralId_body  (optional)
+ * referralId Long ID of referral to update
+ * returns inline_response_204
+ **/
+exports.updateReferral = function(body,referralId) {
+  return new Promise(function(resolve, reject) {
+    var examples = {};
+    examples['application/json'] = {
+  "message" : "message"
+};
     if (Object.keys(examples).length > 0) {
       resolve(examples[Object.keys(examples)[0]]);
     } else {
